cli/sql: move getServerValue to sql_util
